pdf-watcher/requirements.txt
watchdog>=4.0.0
requests>=2.31.0
pdfplumber>=0.11.0
python-dotenv>=1.0.1
